File reorganization for packaging
